refactor: Migrate NewConsultationScreen to MVI
This commit refactors the `NewConsultationScreen` by:
- Deleting the old `NewConsultationScreen.kt` file from the `consultation/screen` package.
- Creating a new `NewConsultationScreen.kt` file in the `consultation/screen/add` package.
- Updating the new `NewConsultationScreen` to adopt the Model-View-Intent (MVI) architecture using `NewConsultationViewModel`.

**Key changes in the new `NewConsultationScreen`:**
- Integrates `NewConsultationViewModel` (using Koin for dependency injection).
- Manages UI state (`ConsultationState`) and handles UI events (`ConsultationEvent`).
- Uses `LaunchedUiEffectHandler` to manage side effects (`ConsultationEffect`) such as navigation and displaying success/error messages.
- Displays success and error messages using `SuccessesDialog` and `DialogWithIcon` respectively.
- Simplifies navigation by using `navigateUp()` and `navigateToMedicalRecord(patientId)` directly.
- Removes manual CoroutineScope and delay management for dialogs, now handled by the `LaunchedUiEffectHandler` and ViewModel.
- Removes direct ViewModel dependencies for `PharmacyViewModel` and `CenterViewModel`, as their data is now managed through `NewConsultationViewModel`.

**Affected files:**
- **features/consultations/src/staff/kotlin/eg/edu/cu/csds/icare/consultation/screen/NewConsultationScreen.kt:** (deleted)
- **features/consultations/src/staff/kotlin/eg/edu/cu/csds/icare/consultation/screen/add/NewConsultationScreen.kt:** (new file)
    - Implemented MVI pattern with `NewConsultationViewModel`.
    - Added `LaunchedUiEffectHandler` for side effects.
    - Updated navigation and dialog handling.
    - Simplified dependencies.
