+++ features/consultations/src/staff/kotlin/eg/edu/cu/csds/icare/consultation/screen/add/NewConsultationScreen.kt
@@ -1,0 +1,171 @@
+package eg.edu.cu.csds.icare.consultation.screen.add
+
+import android.content.Context
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.automirrored.filled.ArrowBack
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.Scaffold
+import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
+import androidx.compose.material3.TopAppBar
+import androidx.compose.material3.TopAppBarDefaults
+import androidx.compose.material3.pulltorefresh.PullToRefreshDefaults.Indicator
+import androidx.compose.material3.pulltorefresh.pullToRefresh
+import androidx.compose.material3.pulltorefresh.rememberPullToRefreshState
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.res.stringResource
+import androidx.constraintlayout.compose.ConstraintLayout
+import androidx.constraintlayout.compose.Dimension
+import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import eg.edu.cu.csds.icare.admin.R
+import eg.edu.cu.csds.icare.consultation.screen.ConsultationDetailsContent
+import eg.edu.cu.csds.icare.consultation.screen.ConsultationEffect
+import eg.edu.cu.csds.icare.consultation.screen.ConsultationEvent
+import eg.edu.cu.csds.icare.core.ui.common.LaunchedUiEffectHandler
+import eg.edu.cu.csds.icare.core.ui.theme.XS_PADDING
+import eg.edu.cu.csds.icare.core.ui.theme.Yellow500
+import eg.edu.cu.csds.icare.core.ui.theme.backgroundColor
+import eg.edu.cu.csds.icare.core.ui.theme.barBackgroundColor
+import eg.edu.cu.csds.icare.core.ui.view.DialogWithIcon
+import eg.edu.cu.csds.icare.core.ui.view.SuccessesDialog
+import kotlinx.coroutines.delay
+import org.koin.androidx.compose.koinViewModel
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+internal fun NewConsultationScreen(
+    viewModel: NewConsultationViewModel = koinViewModel(),
+    navigateUp: () -> Unit,
+    navigateToMedicalRecord: (String) -> Unit,
+) {
+    val context: Context = LocalContext.current
+    val refreshState = rememberPullToRefreshState()
+    val uiState by viewModel.uiState.collectAsStateWithLifecycle()
+    var showSuccessDialog by remember { mutableStateOf(false) }
+    var alertMessage by remember { mutableStateOf("") }
+    var showAlert by remember { mutableStateOf(false) }
+
+    LaunchedUiEffectHandler(
+        viewModel.effect,
+        onConsumeEffect = { viewModel.processEvent(ConsultationEvent.ConsumeEffect) },
+        onEffect = { effect ->
+            when (effect) {
+                is ConsultationEffect.NavigateToMedicalRecord ->
+                    navigateToMedicalRecord(effect.patientId)
+
+                is ConsultationEffect.ShowSuccess -> {
+                    showSuccessDialog = true
+                    delay(timeMillis = 3000)
+                    navigateUp()
+                }
+
+                is ConsultationEffect.ShowError -> {
+                    alertMessage = effect.message.asString(context)
+                    showAlert = true
+                    delay(timeMillis = 3000)
+                    showAlert = false
+                }
+            }
+        },
+    )
+
+    Scaffold(
+        topBar = {
+            TopAppBar(
+                title = { Text(text = stringResource(id = R.string.new_consultation)) },
+                colors =
+                    TopAppBarDefaults.topAppBarColors(
+                        containerColor = barBackgroundColor,
+                        navigationIconContentColor = Color.White,
+                        titleContentColor = Color.White,
+                        actionIconContentColor = Color.White,
+                    ),
+                navigationIcon = {
+                    IconButton(onClick = { navigateUp() }) {
+                        Icon(
+                            imageVector = Icons.AutoMirrored.Filled.ArrowBack,
+                            contentDescription = null,
+                            tint = Color.White,
+                        )
+                    }
+                },
+            )
+        },
+    ) { paddingValues ->
+        Surface(
+            modifier =
+                Modifier
+                    .fillMaxSize()
+                    .pullToRefresh(
+                        state = refreshState,
+                        isRefreshing = uiState.isLoading,
+                        onRefresh = {
+                            viewModel.processEvent(ConsultationEvent.Refresh)
+                        },
+                    ).padding(paddingValues),
+        ) {
+            ConstraintLayout(
+                modifier =
+                    Modifier
+                        .background(backgroundColor)
+                        .fillMaxWidth(),
+            ) {
+                val (refresh, line, content) = createRefs()
+                Box(
+                    modifier =
+                        Modifier
+                            .constrainAs(line) {
+                                top.linkTo(parent.top)
+                                start.linkTo(parent.start)
+                                end.linkTo(parent.end)
+                            }.background(Yellow500)
+                            .fillMaxWidth()
+                            .height(XS_PADDING),
+                )
+
+                ConsultationDetailsContent(
+                    modifier =
+                        Modifier.constrainAs(content) {
+                            top.linkTo(line.bottom)
+                            start.linkTo(parent.start)
+                            end.linkTo(parent.end)
+                            bottom.linkTo(parent.bottom)
+                            width = Dimension.fillToConstraints
+                            height = Dimension.fillToConstraints
+                        },
+                    uiState = uiState,
+                    onEvent = viewModel::processEvent,
+                )
+
+                Indicator(
+                    modifier =
+                        Modifier.constrainAs(refresh) {
+                            top.linkTo(parent.top)
+                            start.linkTo(parent.start)
+                            end.linkTo(parent.end)
+                        },
+                    isRefreshing = uiState.isLoading,
+                    state = refreshState,
+                )
+
+                if (showSuccessDialog) SuccessesDialog {}
+                if (showAlert) DialogWithIcon(text = alertMessage) { showAlert = false }
+            }
+        }
+    }
+}
